Split big file into smaller files
