+++ packages/langium/src/serializer/reassembler/AstDisassembler.ts
@@ -4,161 +4,18 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import type { Reference, RootCstNode } from '../syntax-tree.js';
-import { isRootCstNode, type AstNode, type CstNode, type Mutable, isCompositeCstNode, isLeafCstNode, isAstNode, isReference } from '../syntax-tree.js';
-import { streamAst } from '../utils/ast-utils.js';
-import { streamCst } from '../utils/cst-utils.js';
-import { BiMap, type LangiumCoreServices, assertType, type ParseResult, assertUnreachable } from '../index.js';
-import { type AbstractElement, type Grammar, isAbstractElement } from '../languages/generated/ast.js';
-import type { ILexingError, IRecognitionException, TokenType } from 'chevrotain';
-
-enum InstructionType {
-    Allocate,
-    Element,
-    TokenType,
-    Property,
-    Properties,
-    LinkNode,
-    LinkNodes,
-    Reference,
-    References,
-    Empty,
-    Error,
-    Return
-}
-
-enum NodeType {
-    Cst,
-    Ast
-}
-
-interface AstAssemblerInstructionBase {
-    $type: InstructionType;
-}
-
-type ReferenceInfo = {
-    refText: string;
-    refNode?: number;
-};
-
-enum ErrorSource {
-    Lexer,
-    Parser
-}
-
-namespace Instructions {
-    export interface Allocate extends AstAssemblerInstructionBase {
-        $type: InstructionType.Allocate;
-        cstNodeCount: number;
-        astNodeCount: number;
-    }
-    export interface TokenType extends AstAssemblerInstructionBase {
-        $type: InstructionType.TokenType;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        tokenName: string;
-    }
-    export interface Element extends AstAssemblerInstructionBase {
-        $type: InstructionType.Element;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        value: number;
-    }
-    export interface Property extends AstAssemblerInstructionBase {
-        $type: InstructionType.Property;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        value: number;
-    }
-    export interface Properties extends AstAssemblerInstructionBase {
-        $type: InstructionType.Properties;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        values: Array<number | boolean | string | bigint>;
-    }
-    export interface Reference extends AstAssemblerInstructionBase {
-        $type: InstructionType.Reference;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        refText: string;
-        refNode?: number;
-    }
-    export interface References extends AstAssemblerInstructionBase {
-        $type: InstructionType.References;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-        references: ReferenceInfo[];
-    }
-    export interface LinkNode extends AstAssemblerInstructionBase {
-        $type: InstructionType.LinkNode;
-        sourceKind: NodeType;
-        sourceId: number;
-        targetKind: NodeType;
-        property: string;
-        targetId: number;
-    }
-    export interface LinkNodes extends AstAssemblerInstructionBase {
-        $type: InstructionType.LinkNodes;
-        sourceKind: NodeType;
-        sourceId: number;
-        targetKind: NodeType;
-        property: string;
-        targetIds: number[];
-    }
-    export interface Empty extends AstAssemblerInstructionBase {
-        $type: InstructionType.Empty;
-        sourceKind: NodeType;
-        sourceId: number;
-        property: string;
-    }
-    export interface Error extends AstAssemblerInstructionBase {
-        $type: InstructionType.Error;
-        source: ErrorSource;
-        items: Record<string, unknown>;
-    }
-    export interface Return extends AstAssemblerInstructionBase {
-        $type: InstructionType.Return;
-        rootAstNodeId: number;
-    }
-}
-
-export type AstAssemblerInstruction =
-    | Instructions.Allocate
-    | Instructions.TokenType
-    | Instructions.Property
-    | Instructions.Element
-    | Instructions.Properties
-    | Instructions.Reference
-    | Instructions.References
-    | Instructions.LinkNode
-    | Instructions.LinkNodes
-    | Instructions.Empty
-    | Instructions.Error
-    | Instructions.Return;
+import type { Reference } from '../../syntax-tree.js';
+import { isRootCstNode, type AstNode, type CstNode, type Mutable, isCompositeCstNode, isLeafCstNode, isAstNode, isReference } from '../../syntax-tree.js';
+import { streamAst } from '../../utils/ast-utils.js';
+import { streamCst } from '../../utils/cst-utils.js';
+import { BiMap, type LangiumCoreServices, assertType, type ParseResult } from '../../index.js';
+import { type AbstractElement } from '../../languages/generated/ast.js';
+import type { AstAssemblerInstruction, Instructions, ReferenceData } from './AstAssemblerInstruction.js';
+import { InstructionType, NodeType, ErrorSource } from './AstAssemblerInstruction.js';
+import { createGrammarElementIdMap } from './AstAssemblerInstruction.js';
 
 export interface AstDisassembler {
     disassemble(parseResult: ParseResult<AstNode>): Generator<AstAssemblerInstruction, void, void>;
-}
-
-export interface AstReassemblerContext {
-    lexerErrors: ILexingError[];
-    parserErrors: IRecognitionException[];
-    idToAstNode: Array<Record<string, unknown>>;
-    idToCstNode: Array<Record<string, unknown>>;
-    rootAstNodeId: number;
-    elementToId: BiMap<AbstractElement, number>;
-}
-
-export interface AstReassembler {
-    initializeContext(): AstReassemblerContext;
-    reassemble(context: AstReassemblerContext, instr: AstAssemblerInstruction): boolean;
-    buildParseResult<T extends AstNode>(context: AstReassemblerContext): ParseResult<T>;
 }
 
 export class DefaultAstDisassembler implements AstDisassembler {
@@ -281,9 +138,8 @@ export class DefaultAstDisassembler implements AstDisassembler {
             }
             yield setProperty('offset', node.offset);
             yield setProperty('length', node.length);
-            yield setProperty('end', node.end);
             yield setLink('root', NodeType.Cst, this.cstNodeToId.get(node.root)!);
-            if(node.container) {
+            if (node.container) {
                 yield setLink('container', NodeType.Cst, this.cstNodeToId.get(node.container)!);
             }
             yield setProperty('startLine', node.range.start.line);
@@ -361,7 +217,7 @@ export class DefaultAstDisassembler implements AstDisassembler {
                 return instr;
             };
 
-            const setReferences = (property: string, references: ReferenceInfo[]) => {
+            const setReferences = (property: string, references: ReferenceData[]) => {
                 const instr = <Instructions.References>{
                     $type: InstructionType.References,
                     sourceKind,
@@ -373,7 +229,7 @@ export class DefaultAstDisassembler implements AstDisassembler {
                 return instr;
             };
 
-            const setReference = (property: string, reference: ReferenceInfo) => {
+            const setReference = (property: string, reference: ReferenceData) => {
                 const instr = <Instructions.Reference>{
                     $type: InstructionType.Reference,
                     sourceKind,
@@ -418,7 +274,7 @@ export class DefaultAstDisassembler implements AstDisassembler {
                             yield setLinks(name, NodeType.Ast, value.map(v => this.astNodeToId.get(v)!));
                         } else if (isReference(item)) {
                             assertType<Reference[]>(value);
-                            yield setReferences(name, value.map(v => (<ReferenceInfo>{
+                            yield setReferences(name, value.map(v => (<ReferenceData>{
                                 refText: v.$refText,
                                 refNode: v.$refNode ? this.cstNodeToId.get(v.$refNode) : undefined
                             })));
@@ -432,7 +288,7 @@ export class DefaultAstDisassembler implements AstDisassembler {
                 } else if (isAstNode(value)) {
                     yield setLink(name, NodeType.Ast, this.astNodeToId.get(value)!);
                 } else if (isReference(value)) {
-                    yield setReference(name, <ReferenceInfo>{
+                    yield setReference(name, <ReferenceData>{
                         refText: value.$refText,
                         refNode: value.$refNode ? this.cstNodeToId.get(value.$refNode) : undefined
                     });
@@ -474,157 +330,5 @@ export class DefaultAstDisassembler implements AstDisassembler {
         [...streamAst(astRoot)].forEach((astNode, index) => {
             this.astNodeToId.set(astNode, index);
         });
-    }
-}
-
-function createGrammarElementIdMap(grammar: Grammar) {
-    const result = new BiMap<AbstractElement, number>();
-    let id = 0;
-    for (const element of streamAst(grammar)) {
-        if (isAbstractElement(element)) {
-            result.set(element, id++);
-        }
-    }
-    return result;
-}
-
-class CstClone {
-    root: RootCstNode;
-    offset: number;
-    end: number;
-    get text() {
-        return this.root.fullText.substring(this.offset, this.end);
-    }
-}
-
-export class DefaultAstReassembler implements AstReassembler {
-    private readonly grammarElementIdMap: BiMap<AbstractElement, number>;
-    private readonly grammarTokenTypeIdMap: BiMap<TokenType, string>;
-    constructor(services: LangiumCoreServices) {
-        this.grammarElementIdMap = createGrammarElementIdMap(services.Grammar);
-        const tokens = services.parser.TokenBuilder.buildTokens(services.Grammar) as TokenType[];
-        this.grammarTokenTypeIdMap = new BiMap(tokens.map(tk => [tk, tk.name] as const));
-    }
-
-    buildParseResult<T extends AstNode>(context: AstReassemblerContext): ParseResult<T> {
-        return {
-            lexerErrors: context.lexerErrors,
-            parserErrors: context.parserErrors,
-            value: context.idToAstNode[context.rootAstNodeId] as T
-        };
-    }
-    initializeContext(): AstReassemblerContext {
-        return {
-            rootAstNodeId: -1,
-            idToAstNode: [],
-            idToCstNode: [],
-            lexerErrors: [],
-            parserErrors: [],
-            elementToId: this.grammarElementIdMap
-        };
-    }
-
-    reassemble(ctx: AstReassemblerContext, instr: AstAssemblerInstruction): boolean {
-        switch (instr.$type) {
-            case InstructionType.Allocate:
-                ctx.idToCstNode = Array.from({ length: instr.cstNodeCount }).map(() => (new CstClone() as Mutable<CstNode>));
-                ctx.idToAstNode = Array.from({ length: instr.astNodeCount }).map(() => ({} as Mutable<AstNode>));
-                break;
-            case InstructionType.Empty:
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = [];
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = [];
-                }
-                break;
-            case InstructionType.Element:
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = ctx.elementToId.getKey(instr.value);
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = ctx.elementToId.getKey(instr.value);
-                }
-                break;
-            case InstructionType.Property:
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = instr.value;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = instr.value;
-                }
-                break;
-            case InstructionType.Properties:
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = instr.values;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = instr.values;
-                }
-                break;
-            case InstructionType.Reference: {
-                const reference = <Reference>{
-                    $refText: instr.refText,
-                    $refNode: instr.refNode ? ctx.idToCstNode[instr.refNode] : undefined
-                };
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = reference;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = reference;
-                }
-                break;
-            }
-            case InstructionType.References: {
-                const references = instr.references.map(r => (<Reference>{
-                    $refText: r.refText,
-                    $refNode: r.refNode ? ctx.idToCstNode[r.refNode] : undefined
-                }));
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = references;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = references;
-                }
-                break;
-            }
-            case InstructionType.LinkNode: {
-                const node = instr.targetKind === NodeType.Ast ? ctx.idToAstNode[instr.targetId] : ctx.idToCstNode[instr.targetId];
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = node;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = node;
-                }
-                break;
-            }
-            case InstructionType.LinkNodes: {
-                const nodes = instr.targetKind === NodeType.Ast
-                    ? instr.targetIds.map(id => ctx.idToAstNode[id])
-                    : instr.targetIds.map(id => ctx.idToCstNode[id])
-                    ;
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = nodes;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = nodes;
-                }
-                break;
-            }
-            case InstructionType.Return:
-                ctx.rootAstNodeId = instr.rootAstNodeId;
-                return true;
-            case InstructionType.Error:
-                if (instr.source === ErrorSource.Lexer) {
-                    ctx.lexerErrors.push({ ...instr.items } as unknown as ILexingError);
-                } else {
-                    ctx.parserErrors.push({ ...instr.items } as unknown as IRecognitionException);
-                }
-                break;
-            case InstructionType.TokenType: {
-                const tokenType = this.grammarTokenTypeIdMap.getKey(instr.tokenName)!;
-                if (instr.sourceKind === NodeType.Ast) {
-                    ctx.idToAstNode[instr.sourceId][instr.property] = tokenType;
-                } else {
-                    ctx.idToCstNode[instr.sourceId][instr.property] = tokenType;
-                }
-                break;
-            }
-            default:
-                assertUnreachable(instr);
-        }
-        return false;
     }
 }
