+++ packages/langium/src/serializer/reassembler/index.ts
@@ -3,6 +3,6 @@
  * This program and the accompanying materials are made available under the
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
-export * from './reassembler/index.js';
-export * from './hydrator.js';
-export * from './json-serializer.js';
+export * from './AstAssemblerInstruction.js';
+export * from './AstDisassembler.js';
+export * from './AstReassembler.js';
