+++ packages/langium/src/services.ts
@@ -37,7 +37,7 @@ import type { IndexManager } from './workspace/index-manager.js';
 import type { WorkspaceLock } from './workspace/workspace-lock.js';
 import type { Hydrator } from './serializer/hydrator.js';
 import type { WorkspaceManager } from './workspace/workspace-manager.js';
-import type { AstDisassembler, AstReassembler } from './serializer/beamer.js';
+import type { AstDisassembler, AstReassembler } from './serializer/reassembler/index.js';
 
 /**
  * The services generated by `langium-cli` for a specific language. These are derived from the
